Include UUID generator.
In the refactoring objects will be identified with uuid's. This allows to get rid of the factory pattern used to create tasks. It furthermore allows to save tasks and board in separate files. That would make it possible to use the plugin in a team and contribute tasks to git as own files.

The generator is taken from binogure-studio/godot-uuid.