Task cards also reused (f907c41 was about reusing stage)
Task view "cache" also added to board. (Meanwhile stage "cache" was also fixed and tidied)
If a stage has a parent_board_view set, it requests task view from it.
In this way task card are also reused. (Before this, e.g., a task card forgot whether its step list was expanded or not during drag-and-drop)